more pkg-moving
