+++ src/main/scala/magic_rdds/size.scala
@@ -1,0 +1,6 @@
+package magic_rdds
+
+import org.hammerlab.magic.rdd
+
+trait size extends rdd.size
+object size extends size
