+++ src/main/scala/magic_rdds/sort.scala
@@ -1,0 +1,7 @@
+package magic_rdds
+
+import org.hammerlab.magic.rdd
+
+trait sort extends rdd.sort
+object sort extends sort
+
