--- src/main/scala/org/hammerlab/magic/rdd/package.scala
@@ -1,4 +1,0 @@
-package org.hammerlab.magic
-
-package object rdd
-  extends sliding.ops
