--- src/main/scala/org/hammerlab/magic/rdd/size/package.scala
@@ -1,3 +1,0 @@
-package org.hammerlab.magic.rdd
-
-package object size extends Size
