--- src/main/scala/org/hammerlab/magic/rdd/sliding/ops.scala
@@ -1,7 +1,0 @@
-package org.hammerlab.magic.rdd.sliding
-
-trait ops
-  extends Sliding
-  with BorrowElems
-
-object ops extends ops
